third_party/WebKit/LayoutTests/http/tests/storage/callbacks-are-called-in-correct-context.html

<html>
<head>
<script>
function setupAndRunFramesTest()
{
    if (window.testRunner) {
        testRunner.dumpAsText();
        testRunner.dumpChildFrames();
        testRunner.waitUntilDone();
    }
    top.frames[0].document.getElementById("console").innerText = "";
    top.frames[1].runTest();
}
</script>
</head>
<frameset onload="setupAndRunFramesTest()">
<frame name="outputFrame" src="resources/callbacks-are-called-in-correct-context-output-frame.html">
<frame name="firstFrame" src="resources/callbacks-are-called-in-correct-context-first-frame.html">
<frame name="secondFrame" src="resources/callbacks-are-called-in-correct-context-second-frame.html">
<frame name="thirdFrame" src="resources/callbacks-are-called-in-correct-context-third-frame.html">
</frameset>
</html>
